test(server): split parquet.rs tests into ratchet-exempt sibling file
The finish_combined() copy-elimination pushed parquet.rs to 702 lines, past
its 584-line module-size budget. Move the inline `#[cfg(test)] mod tests`
verbatim into a sibling `parquet_tests.rs`, declared from parquet.rs via
`#[cfg(test)] #[path = "parquet_tests.rs"] mod parquet_tests;`. The `#[path]`
keeps the tests a CHILD module of parquet (so `use super::*` still reaches
the file's private helpers) while the file lives beside parquet.rs; test
files are ratchet-exempt. parquet.rs drops to 522 lines. Ratchet budget row
lowered 584 -> 522 (ratchet down, never up). No production logic changed.
